### 2023-02-21 #82 @teddygroves
bibat: Batteries-included Bayesian analysis template
Submitting Author: Teddy Groves (@teddygroves)
Package Name: bibat
One-Line Description of Package: a batteries-included template for Bayesian data analysis projects
Repository Link (if existing):  https://github.com/teddygroves/bibat

---

Bibat is a Python package providing a flexible interactive template for Bayesian statistical analysis projects.

It aims to make it easier to create software projects that implement a Bayesian workflow that scales to arbitrarily many inter-related statistical models, data transformations, inferences and computations. Bibat also aims to promote software quality by providing a modular, automated and reproducible project that takes advantage of and integrates together the most up to date statistical software.

Bibat comes with "batteries included" in the sense that it creates a working example project, which the user can adapt so that it implements their desired analysis. We believe this style of template makes for better usability and easier testing of Bayesian workflow projects compared with the alternative approach of providing an incomplete skeleton project.

I think the best category would be 'workflow automation and versioning' as described [here](https://github.com/pyOpenSci/software-peer-review/pull/162/) but 'reproducibility' works as well!

Bibat makes Bayesian statistical analysis projects easier to reproduce by providing easy access to automation, a file-based and declarative workflow and a project structure with well-designed abstractions and modules.

The target audience is people who want to write software implementing a Bayesian workflow as described in [this paper](https://arxiv.org/abs/2011.01808v1) and are willing to do so using Python scientific libraries, Stan, cmdstanpy, pydantic, pandera and make, as well as perhaps pytest, sphinx, quarto and github actions.

I am not aware of any interactive template that specifically targets a Python-based Bayesian workflow that scales to arbitrarily many models and data transformations.

In addition, bibat is unusual compared to other interactive templates because it is 'batteries-inclued', providing a full working example project rather than an incomplete skeleton

I previously made a presubission enquiry (#62) about an earlier version of this package, which pyOpenSci could not review at the time.

I am making another enquiry to see if it is now in scope, considering the latest state of the package and pyOpenSci's latest policy and capacity.

The following changes might be particularly relevant:

- bibat is now a Python package rather than a cookiecutter template.
- bibat's documentation is now much more comprehensive, including a detailed [vignette](https://bibat.readthedocs.io/en/latest/_static/report.html) explaining how to implement a complex statistical analysis.
- bibat would now avoid being flagged by joss for having fewer than 1000 lines of non-notebook Python or R code (see [recent cloc action here](https://github.com/teddygroves/bibat/actions/runs/4232874413/jobs/7353135139)).

I also noticed that a broadly similar package was recently considered within scope (#74).

- [x] Please make sure that you took notice of our [Community Code of Conduct][Conduct] and are able to commit to the package maintainance as per our [Policies Guidelines][Commitment].

[Conduct]: https://www.pyopensci.org/peer-review-guide/about-peer-review/code-of-conduct.html
