### 2023-02-24 #82 @NickleDave
Hi @teddygroves! Just want to let you know we saw this presubmission

It's obvious you've put a ton of work into `bibat`.
I do see you've made changes that meet many of our requirements for scope and the requirements for JOSS articles, as you describe.

We are discussing with people more familiar with Bayesian workflows and STAN. This presubmission inquiry also brings up some questions about how we define the category of reproducibility that we need to answer.

Please expect an answer from us sometime next week. Thank you!
