### 2023-02-28 7812854
Use find_namespace: in setup.cfg instead of __init__.py
