+++ setup.cfg
@@ -31,7 +31,12 @@ install_requires =
     click
     pydantic
 python_requires = >=3.10
-packages = bibat
+packages = find_namespace:
+package_dir =
+    =bibat
+
+[options.packages.find]
+where = bibat
 
 [options.entry_points]
 console_scripts =
