### 2023-02-28 407daff
Add extra inference folders to baseball example
