+++ setup.cfg
@@ -31,12 +31,10 @@ install_requires =
     click
     pydantic
 python_requires = >=3.10
-packages = find_namespace:
-package_dir =
-    =bibat
+packages =
+    bibat
+    bibat.hooks
 
-[options.packages.find]
-where = bibat
 
 [options.entry_points]
 console_scripts =
