+++ setup.cfg
@@ -4,7 +4,7 @@ url = https://github.com/teddygroves/mbag
 download_url = https://pypi.org/project/mbag
 author = Teddy Groves
 author_email = tedgro@dtu.dk
-version = 0.0.12
+version = 0.0.13
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Science/Research
