### 2023-02-28 #82 @NickleDave
Hi @teddygroves we have decided to go ahead with a review.

Some context, to keep thing open: the general consensus is that this package clearly fits into the replication category, and that it would be considered in scope by our sister org rOpenSci, along the lines of tools for [research compendia](https://the-turing-way.netlify.app/reproducible-research/compendia.html#further-reading) AKA [repro-packs](https://lorenabarba.com/blog/how-repro-packs-can-save-your-future-self/). Like I said before, it's clear you've put a ton of work in to development. There's a question of how generalized the approach is, or whether the community you're developing for might see it as "opinionated", but a good way to find out would be to put the package through review!

Please go ahead and make a full submission. Once you do so, I will close this issue. I'll proceed with finding an editor in the meantime.
