### 2023-03-01 e192e57
Update version and fix incorrect url in setup.cfg
