+++ setup.cfg
@@ -1,10 +1,10 @@
 [metadata]
 name = bibat
-url = https://github.com/teddygroves/mbag
-download_url = https://pypi.org/project/mbag
+url = https://github.com/teddygroves/bibat
+download_url = https://pypi.org/project/bibat
 author = Teddy Groves
 author_email = tedgro@dtu.dk
-version = 0.0.13
+version = 0.1.0
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Science/Research
@@ -18,7 +18,7 @@ classifiers =
     Programming Language :: Python :: 3 :: Only
     Topic :: Scientific/Engineering :: Information Analysis
 license = GNU General Public License version 3
-description = Medium-sized Bayesian Analysis Generator
+description = Batteries-included Bayesian analysis template
 long_description = file: README.rst
 keywords =
 
