+++ tox.ini
@@ -1,25 +1,26 @@
-[flake8]
-exclude = *.venv*, *.tox*, build/*
-max_line_length = 80
-
 [tox]
 env_list =
     format
     flake8
-    py311
+    py3{9, 10, 11}
 minversion = 4.2.8
+
+[flake8]
+exclude = *.venv*, *.tox*, build/*
+max_line_length = 80
 
 [testenv]
 description = run the tests with pytest
 package = wheel
 wheel_build_env = .pkg
 deps =
-    pytest>=6
+    pytest>=7
+    pytest-sugar
 commands =
-    pytest {tty:--color=yes} {posargs}
+    pytest {tty:--color=yes} tests
 
 [testenv:format]
-description = install black in a virtual environment and invoke it on the current folder
+description = install black and isort and invoke them on the current folder
 deps =
     black[jupyter]
     isort
@@ -27,13 +28,6 @@ skip_install = true
 commands =
     black --check --diff .
     isort --check-only --diff .
-
-[testenv:py311]
-description = install pytest in a virtual environment and invoke it on the tests folder
-deps =
-    pytest>=7
-    pytest-sugar
-commands = pytest tests {posargs}
 
 [testenv:flake8]
 skip_install = True
