+++ setup.cfg
@@ -13,6 +13,7 @@ classifiers =
     Operating System :: MacOS :: MacOS X
     Operating System :: Microsoft :: Windows :: Windows 10
     Operating System :: POSIX
+    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3 :: Only
@@ -30,7 +31,7 @@ install_requires =
     cookiecutter
     click
     pydantic
-python_requires = >=3.10
+python_requires = >=3.9
 packages =
     bibat
     bibat.hooks
