+++ setup.cfg
@@ -4,7 +4,7 @@ url = https://github.com/teddygroves/bibat
 download_url = https://pypi.org/project/bibat
 author = Teddy Groves
 author_email = tedgro@dtu.dk
-version = 0.1.0
+version = 0.1.1
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Science/Research
