+++ Makefile
@@ -1,0 +1,8 @@
+.PHONY: qa
+
+## Apply code quality assurance tools.
+qa:
+	black .
+	isort .
+
+
