### 2023-03-27 6c08b90
Fail before installation for unsupported Pythons
