### 2023-03-27 85bb4c4
Update docstrings and add automatic api docs
