### 2023-03-27 131fefa
Bump version and fix entry point
