+++ setup.cfg
@@ -4,7 +4,7 @@ url = https://github.com/teddygroves/bibat
 download_url = https://pypi.org/project/bibat
 author = Teddy Groves
 author_email = tedgro@dtu.dk
-version = 0.1.1
+version = 0.1.2
 classifiers =
     Development Status :: 2 - Pre-Alpha
     Intended Audience :: Science/Research
@@ -39,7 +39,7 @@ packages =
 
 [options.entry_points]
 console_scripts =
-    bibat = bibat.cli:main
+    bibat = bibat.cli:generate_project
 
 [options.extras_require]
 development =
