### 2023-03-27 e89ea69
add sphinx-click to docs requirements
