+++ docs/requirements.txt
@@ -1,1 +1,2 @@
 pydata-sphinx-theme
+sphinx-click
