### 2023-04-27 737cc8c
tidy docs and add readthedocs config file
