### 2023-04-27 eb79c66
configure readthedocs to install the package
