+++ .readthedocs.yaml
@@ -22,5 +22,9 @@ sphinx:
 # Optionally declare the Python requirements required to build your docs
 python:
    install:
-   - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+      extra_requirements:
+        - development
+    - requirements: docs/requirements.txt
 
