### 2023-04-27 4e2a687
tweaking .readthedocs.yaml
