+++ .readthedocs.yaml
@@ -12,8 +12,9 @@ build:
     python: "3.11"
 
 # Build documentation in the docs/ directory with Sphinx
-sphinx:
-   configuration: docs/conf.py
+sphinx: 
+  configuration: conf.py
+  fail_on_warning: true
 
 # If using Sphinx, optionally build your docs in additional formats such as PDF
 # formats:
@@ -26,5 +27,7 @@ python:
       path: .
       extra_requirements:
         - development
+    - method: pip
+      path: bibat   
     - requirements: docs/requirements.txt
-
+   system_packages: true
