### 2023-04-27 dbe8e18
tweaking .readthedocs.yaml
