### 2023-04-27 645e644
Add bibtex citation info to readme
