### 2023-04-27 9a6a364
switch to pyproject.toml only, removing setup.cfg and setup.py
