+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2022, Teddy Groves"
 author = "Teddy Groves"
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.2"
+release = "0.1.3"
 
 root_doc = "index"
 
