+++ setup.cfg
@@ -1,55 +1,0 @@
-[metadata]
-name = bibat
-url = https://github.com/teddygroves/bibat
-download_url = https://pypi.org/project/bibat
-author = Teddy Groves
-author_email = tedgro@dtu.dk
-version = 0.1.2
-classifiers =
-    Development Status :: 2 - Pre-Alpha
-    Intended Audience :: Science/Research
-    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
-    Natural Language :: English
-    Operating System :: MacOS :: MacOS X
-    Operating System :: Microsoft :: Windows :: Windows 10
-    Operating System :: POSIX
-    Programming Language :: Python :: 3.9
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
-    Programming Language :: Python :: 3 :: Only
-    Topic :: Scientific/Engineering :: Information Analysis
-license = GNU General Public License version 3
-description = Batteries-included Bayesian analysis template
-long_description = file: README.rst
-keywords =
-
-[options]
-zip_safe = True
-include_package_data = True
-install_requires =
-    pip >= 20
-    cookiecutter
-    click
-    pydantic
-python_requires = >=3.9
-packages =
-    bibat
-    bibat.hooks
-
-
-[options.entry_points]
-console_scripts =
-    bibat = bibat.cli:generate_project
-
-[options.extras_require]
-development =
-    black
-    isort
-    pytest
-    tox
-    sphinx
-    sphinx-click
-    pydata_sphinx_theme
-
-[bdist_wheel]
-universal = 1
