+++ pyproject.toml
@@ -2,6 +2,55 @@
 requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
+[project]
+name = "bibat"
+authors = [
+    {name = "Teddy Groves", email = "tedgro@dtu.dk"},
+]
+version = "0.1.3"
+classifiers = [
+    "Development Status :: 2 - Pre-Alpha",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
+    "Natural Language :: English",
+    "Operating System :: MacOS :: MacOS X",
+    "Operating System :: Microsoft :: Windows :: Windows 10",
+    "Operating System :: POSIX",
+    "Topic :: Scientific/Engineering :: Information Analysis",
+]
+license = {text = "GNU General Public License version 3"}
+description = "Batteries-included Bayesian analysis template"
+readme = "README.rst"
+requires-python = ">=3.9"
+dependencies = [
+    "pip >= 20",
+    "cookiecutter",
+    "click",
+    "pydantic",
+]
+
+[project.optional-dependencies]
+development = [
+    "black",
+    "isort",
+    "pytest",
+    "tox",
+    "sphinx",
+    "sphinx-click",
+    "pydata_sphinx_theme",
+]
+[project.urls]
+homepage = "https://github.com/teddygroves/bibat"
+download = "https://pypi.org/project/bibat"
+documentation = "https://bibat.readthedocs.io/"
+
+[tool.setuptools]
+include-package-data = true
+packages = ["bibat", "bibat.hooks"]
+
+[project.entry-points.console_scripts]
+bibat = "bibat.cli:generate_project"
+
 [tool.black]
 line-length = 80
 target-version = ['py311']
