+++ bibat/{{cookiecutter.repo_name}}/requirements-tooling.txt
@@ -1,10 +1,0 @@
-pandas-stubs
-types-toml
-flake8
-flake8-bugbear
-flake8-docstrings
-mypy
-python-lsp-server[all]
-python-lsp-black
-pylsp-mypy
-pyls-isort
