### 2023-04-27 1c16c7a
Consolidate data preparation and flatten file structure
