+++ bibat/cli.py
@@ -26,7 +26,7 @@ WIZARD_FIELDS = [
         .replace(" ", "_"),
     ),
     WizardStr("author_name", "What is your name?", default="Author name"),
-    WizardStr("author_email", "What is your email?", default="Author email"),
+    WizardStr("author_email", "What is your email?", default="author@email.com"),
     WizardStr(
         "coc_contact",
         "Who should be the code of conduct contact?",
@@ -85,6 +85,4 @@ def generate_project(config_file: Optional[str]):
     for wizard_field in WIZARD_FIELDS:
         context[wizard_field.name] = prompt_user(wizard_field, context)
     context["bibat_version"] = version("bibat")
-    cookiecutter(
-        THIS_DIR, no_input=True, extra_context=context, config_file=config_file
-    )
+    cookiecutter(THIS_DIR, no_input=True, extra_context=context)
