+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/prepare_data.py
@@ -1,0 +1,6 @@
+"""Run the prepare_data function from the data_preparaion module."""
+
+from {{cookiecutter.repo_name}}.data_preparation import prepare_data
+
+if __name__ == "__main__":
+    prepare_data()
