+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/data_preparation.py
@@ -1,0 +1,178 @@
+"""Provides functions prepare_data_x.
+
+These functions should take in a dataframe of measurements and return a
+PreparedData object.
+
+"""
+import json
+import os
+
+import numpy as np
+import pandas as pd
+import pandera as pa
+from pandera.typing import Series, DataFrame
+from pydantic.dataclasses import dataclass
+from {{cookiecutter.repo_name}}.util import CoordDict, make_columns_lower_case
+
+
+NAME_FILE = "name.txt"
+COORDS_FILE = "coords.json"
+MEASUREMENTS_FILE = "measurements.csv"
+NEW_COLNAMES = {"yButIThoughtIdAddSomeLetters": "y"}
+DROPNA_COLS = ["y"]
+N_CV_FOLDS = 10
+DIMS = {
+    "b": ["covariate"],
+    "y": ["observation"],
+    "yrep": ["observation"],
+    "llik": ["observation"],
+}
+
+HERE = os.path.dirname(__file__)
+DATA_DIR = os.path.join(HERE, "..", "data")
+RAW_DIR = os.path.join(DATA_DIR, "raw")
+PREPARED_DIR = os.path.join(DATA_DIR, "prepared")
+RAW_DATA_FILES = {
+    "raw_measurements": os.path.join(RAW_DIR, "raw_measurements.csv"),
+}
+
+
+def prepare_data():
+    """Main function."""
+    print("Reading raw data...")
+    raw_data = {
+        k: pd.read_csv(v, index_col=None) for k, v in RAW_DATA_FILES.items()
+    }
+    data_preparation_functions_to_run = [
+        prepare_data_interaction,
+        prepare_data_no_interaction,
+        prepare_data_fake_interaction
+    ]
+    print("Preparing data...")
+    for dpf in data_preparation_functions_to_run:
+        print(f"Running data preparation function {dpf.__name__}...")
+        prepared_data = dpf(raw_data["raw_measurements"])
+        output_dir = os.path.join(PREPARED_DIR, prepared_data.name)
+        print(f"\twriting files to {output_dir}")
+        if not os.path.exists(PREPARED_DIR):
+            os.mkdir(PREPARED_DIR)
+        write_prepared_data(prepared_data, output_dir)
+
+
+class MeasurementsDF(pa.SchemaModel):
+    """A PreparedData should have a measurements dataframe like this.
+
+    Other columns are also allowed!
+    """
+
+    x1: Series[float]
+    x2: Series[float]
+    x1colonx2: Series[float] = pa.Field(alias="x1:x2")
+    y: Series[float]
+
+
+@dataclass
+class PreparedData:
+    """What prepared data looks like in this analysis."""
+
+    name: str
+    coords: CoordDict
+    measurements: DataFrame[MeasurementsDF]
+
+
+def load_prepared_data(directory: str) -> PreparedData:
+    """Load prepared data from files in directory."""
+    with open(os.path.join(directory, COORDS_FILE), "r") as f:
+        coords = json.load(f)
+    with open(os.path.join(directory, NAME_FILE), "r") as f:
+        name = f.read()
+    measurements = pd.read_csv(os.path.join(directory, MEASUREMENTS_FILE))
+    return PreparedData(
+        name=name,
+        coords=coords,
+        measurements=measurements,
+    )
+
+
+def write_prepared_data(prepped: PreparedData, directory):
+    """Write prepared data files to a directory."""
+    if not os.path.exists(directory):
+        os.mkdir(directory)
+        prepped.measurements.to_csv(os.path.join(directory, MEASUREMENTS_FILE))
+    with open(os.path.join(directory, COORDS_FILE), "w") as f:
[…]
