+++ bibat/examples/baseball/baseball/data_preparation.py
@@ -1,0 +1,210 @@
+"""Provides functions prepare_data_x.
+
+These functions should take in a dataframe of measurements and return a
+PreparedData object.
+
+"""
+import json
+import os
+
+import numpy as np
+import pandas as pd
+import pandera as pa
+from pandera.typing import DataFrame, Series
+from pydantic.dataclasses import dataclass
+from baseball.util import CoordDict, make_columns_lower_case
+
+NAME_FILE = "name.txt"
+COORDS_FILE = "coords.json"
+MEASUREMENTS_FILE = "measurements.csv"
+NEW_COLNAMES = {"yButIThoughtIdAddSomeLetters": "y"}
+DROPNA_COLS = ["y"]
+N_CV_FOLDS = 10
+DIMS = {
+    "b": ["covariate"],
+    "y": ["observation"],
+    "yrep": ["observation"],
+    "llik": ["observation"],
+}
+
+HERE = os.path.dirname(__file__)
+DATA_DIR = os.path.join(HERE, "..", "data")
+RAW_DIR = os.path.join(DATA_DIR, "raw")
+PREPARED_DIR = os.path.join(DATA_DIR, "prepared")
+RAW_DATA_FILES = {
+    "2006": [os.path.join(RAW_DIR, "2006.csv")],
+    "bdb": [
+        os.path.join(RAW_DIR, "bdb-main.csv"),
+        os.path.join(RAW_DIR, "bdb-post.csv"),
+        os.path.join(RAW_DIR, "bdb-apps.csv"),
+    ],
+}
+
+
+def prepare_data():
+    """Main function."""
+    print("Reading raw data...")
+    raw_data = {
+        k: [pd.read_csv(file, index_col=None) for file in v] 
+        for k, v in RAW_DATA_FILES.items()
+    }
+    data_preparation_functions_to_run = {
+        "2006": prepare_data_2006, "bdb": prepare_data_bdb
+    }
+    print("Preparing data...")
+    for name, dpf in data_preparation_functions_to_run.items():
+        print(f"Running data preparation function {dpf.__name__}...")
+        prepared_data = dpf(*raw_data[name])
+        output_dir = os.path.join(PREPARED_DIR, prepared_data.name)
+        print(f"\twriting files to {output_dir}")
+        if not os.path.exists(PREPARED_DIR):
+            os.mkdir(PREPARED_DIR)
+        write_prepared_data(prepared_data, output_dir)
+
+
+class MeasurementsDF(pa.SchemaModel):
+    """A PreparedData should have a measurements dataframe like this.
+
+    Other columns are also allowed!
+    """
+
+    player_season: Series[str]
+    season: Series[str] = pa.Field(coerce=True)
+    n_attempt: Series[int] = pa.Field(ge=1)
+    n_success: Series[int] = pa.Field(ge=0)
+
+
+@dataclass
+class PreparedData:
+    """What prepared data looks like in this analysis."""
+
+    name: str
+    coords: CoordDict
+    measurements: DataFrame[MeasurementsDF]
+
+
+def load_prepared_data(directory: str) -> PreparedData:
+    """Load prepared data from files in directory."""
+    with open(os.path.join(directory, COORDS_FILE), "r") as f:
+        coords = json.load(f)
+    with open(os.path.join(directory, NAME_FILE), "r") as f:
+        name = f.read()
+    measurements = pd.read_csv(os.path.join(directory, MEASUREMENTS_FILE))
+    return PreparedData(
+        name=name,
+        coords=coords,
+        measurements=measurements,
+    )
+
+
+def write_prepared_data(prepped: PreparedData, directory):
+    """Write prepared data files to a directory."""
+    if not os.path.exists(directory):
[…]
