+++ bibat/examples/baseball/prepare_data.py
@@ -1,46 +1,0 @@
-"""Read the data in RAW_DIR and save prepared data to PREPARED_DIR."""
-
-import os
-
-import pandas as pd
-from src import data_preparation_functions
-from src.prepared_data import write_prepared_data
-
-DATA_PREPARATION_FUNCTIONS_TO_RUN = {
-    "2006": data_preparation_functions.prepare_data_2006,
-    "bdb": data_preparation_functions.prepare_data_bdb,
-}
-
-DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
-RAW_DIR = os.path.join(DATA_DIR, "raw")
-PREPARED_DIR = os.path.join(DATA_DIR, "prepared")
-RAW_DATA_FILES = {
-    "2006": [os.path.join(RAW_DIR, "2006.csv")],
-    "bdb": [
-        os.path.join(RAW_DIR, "bdb-main.csv"),
-        os.path.join(RAW_DIR, "bdb-post.csv"),
-        os.path.join(RAW_DIR, "bdb-apps.csv"),
-    ],
-}
-
-
-def main():
-    """Save prepared data in the PREPARED_DIR."""
-    print("Reading raw data...")
-    raw_data = {
-        k: [pd.read_csv(f, index_col=None) for f in v]
-        for k, v in RAW_DATA_FILES.items()
-    }
-    print("Preparing data...")
-    for name, dpf in DATA_PREPARATION_FUNCTIONS_TO_RUN.items():
-        print(f"Running data preparation function {dpf.__name__}...")
-        prepared_data = dpf(*raw_data[name])
-        output_dir = os.path.join(PREPARED_DIR, prepared_data.name)
-        print(f"\twriting files to {output_dir}")
-        if not os.path.exists(PREPARED_DIR):
-            os.mkdir(PREPARED_DIR)
-        write_prepared_data(prepared_data, output_dir)
-
-
-if __name__ == "__main__":
-    main()

[…]
