+++ bibat/examples/baseball/baseball/inference_configuration.py
@@ -4,8 +4,8 @@ import os
 from typing import Callable, Dict, List, Optional
 
 import toml
-from pydantic import BaseModel, Field, root_validator, validator
 from baseball import stan_input_functions
+from pydantic import BaseModel, Field, root_validator, validator
 
 AVAILABLE_MODES = ["prior", "posterior", "kfold"]
 HERE = os.path.dirname(os.path.abspath(__file__))
