### 2023-04-27 a0cc993
add show_console to sample arguments for github actions debugging - remove later
