+++ bibat/{{cookiecutter.repo_name}}/inferences/no_interaction/config.toml
@@ -17,6 +17,7 @@ warn-pedantic = true
 save_warmup = false
 iter_warmup = 2000
 iter_sampling = 2000
+show_console = true
 
 [sample_kwargs.kfold]
 chains = 1
