### 2023-04-27 e27b3b7
Revert "add show_console to sample arguments for github actions debugging - remove later"
This reverts commit a0cc99377a4a8daf6d24f05a52a8dc4f711e614e.
