+++ .github/workflows/test_end_to_end.yml
@@ -15,10 +15,10 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.11.2
+    - name: Set up Python 3.11.3
       uses: actions/setup-python@v2
       with:
-        python-version: 3.11.2
+        python-version: 3.11.3
     - name: Install bibat
       run: pip install -e .
     - name: Use bibat
