+++ .github/workflows/test_sphinx.yml
@@ -12,10 +12,10 @@ jobs:
     steps:
 
     - uses: actions/checkout@v2
-    - name: Set up Python 3.11.2
+    - name: Set up Python 3.11.3
       uses: actions/setup-python@v2
       with:
-        python-version: 3.11.2
+        python-version: 3.11.3
     - name: install bibat with development packages
       run: pip install -e .\[development\]
     - name: build docs with sphinx
