+++ .github/workflows/run_tox.yml
@@ -15,10 +15,10 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.11.2
+    - name: Set up Python 3.11.3
       uses: actions/setup-python@v2
       with:
-        python-version: 3.11.2
+        python-version: 3.11.3
     - name: Install bibat
       run: |
         python -m pip install --upgrade pip
