### 2023-04-27 fd8b9cc
Link to code of conduct in docs and CONTRIBUTING.md
