### 2023-04-28 28bed9f
Add development installation instructions to readme and docs
