+++ docs/index.rst
@@ -74,10 +74,10 @@ bibat like this:
 
     $ bibat --config_file=my_json_file.json
 
-See `here
+See `bibat's cookiecutter schema
 <https://github.com/teddygroves/bibat/blob/main/bibat/cookiecutter.json>`_ for
-the information you need to create a suitable json file.
-You can use bibat to start your statistical analysis project like this:
+the information you need to create a suitable json file. You can use bibat to
+start your statistical analysis project like this:
 
 .. _intended workflow:
 
