+++ Makefile
@@ -1,8 +1,0 @@
-.PHONY: qa
-
-## Apply code quality assurance tools.
-qa:
-	black .
-	isort .
-
-
