+++ docs/working_with_bibat.rst
@@ -1,0 +1,279 @@
+.. _working_with_bibat:
+
+==================
+Working with bibat
+==================
+
+This page explains how to create custom statistical analyses using bibat,
+including the :ref:`intended workflow <intended_workflow>`, instructions for
+:ref:`editing the example analysis <editing_the_example_analysis>` and links to
+some :ref:`vignettes`.
+
+.. _intended_workflow:
+
+Intended workflow
+=================
+
+bibat assumes that a statistical analysis consists of the following components:
+
+- **Raw data**
+- **Data preparation** in which raw data is transformed to produce prepared data
+  that fit a common structure.
+- **Statistical models**
+- **Inferences**, i.e. combinations of a prepared dataset, a statistical model,
+  a choice of which modes to fit the model in (for example
+  prior mode, posterior mode and exact k-fold cross-validation mode) and how
+  exactly to do so, and the results of doing this fitting.
+- **Investigations** that do things with inferences and prepared data, such as
+  making plots.
+- **Documentation**
+
+These terms are all pretty standard, except "Inference", which has a more
+specific than normal meaning in this context. This choice was made to avoid
+creating new jargon, and in order to match the arviz concept of `InferenceData
+<https://arviz-devs.github.io/arviz/api/inference_data.html>`_.
+
+To perform a statistical analysis means running some data preparation
+operations, then carrying out a set of inferences, doing some investigations
+and providing some documentation. bibat provides functionality for conveniently
+doing these things, as well as explicit and convenient representations of all
+the components and even a ready-made working example project.
+
+bibat's strategy for representing the components of a statistical analysis is as
+follows:
+
+- Raw data are files that live in the directory :code:`data/raw`.
+
+- Source code lives in a folder with the same name as the analysis, for example
+  `my_analysis`.
+
+- Logic for data preparation, including a common structure given by a class
+  PreparedData and functions producing data in this format, lives in the file
+  `my_analysis/data_preparation.py`.
+
+- Statistical models are represented by Stan programs in the directory
+  :code:`my_analysis/stan/`
+
+- Inferences are subdirectories of the directory :code:`inferences`. Each such
+  subdirectory contains a file called :code:`config.toml` that chooses a
+  prepared dataset, statistical model and some fitting modes, as well as sampler
+  configuraiton. When the analysis is run the folder is populated with an
+  :code:`InferenceData` object saved in the file :code:`idata.json`, containing
+  samples and debug information.
+
+- Investigations are performed literately using the notebook file
+  :code:`my_analysis/investigate.ipynb`, which saves plots to the directory
+  :code:`plots`.
+
+- Documentation lives in the directory :code:`docs`, and can be written using a
+  range of standard formats including `quarto <https://quarto.org/>`_ and
+  `Sphinx <https://www.sphinx-doc.org/en/master/>`_.
+
+The analysis is performed by setting up a suitable programming environment and
+then running the Python scripts :code:`my_analysis/prepare_data.py` and
+:code:`my_analysis/sample.py`, which live in the project root, executing the
+notebook file :code:`my_analysis/investigate.ipynb` and building the
+documentation. These tasks are automated using the makefile :code:`Makefile`,
+so that the entire analysis can be performed using the command :code:`make
+analysis` while avoiding unnecessarily re-running any tasks.
+
+After running the analysis, the next step is to make some changes and run a new
+analysis. This can be done by editing any of the files representing the
[…]
