+++ docs/index.rst
@@ -8,29 +8,34 @@ uses `cookiecutter <https://cookiecutter.readthedocs.io/en/latest/>`_, `Stan
 `arviz <https://arviz-devs.github.io/arviz/>`_. It takes a 'batteries-included'
 approach, creating a whole working project rather than a skeleton.
 
-The :ref:`installation` and :ref:`usage` sections explain how to install bibat
-and use it to easily create a folder containing a complete example Bayesian
-statistical analysis project.
+This page has the information you need in order to :ref:`get started
+<getting_started>` with bibat by installing it and then using it to create a
+folder containing a complete example Bayesian statistical analysis project.
 
-The next thing you will probably want to do is edit the example analysis so that
-it does something different. To find out how to do that, get acquainted with
-bibat's :ref:`intended workflow`, check out the section on :ref:`editing the
-example analysis` or explore some of the :ref:`vignettes`.
+The next thing you will probably want to do is learn how to turn the example
+analysis into the analysis you actually want to do. For this there is the
+:ref:`working with bibat <working_with_bibat>` page.
 
-But wait, why did you do all of that? What was the point of this so-called
-"batteries-included template" anyway? The :ref:`motivation` section attempts to
-explain.
-
-If you would like to contribute to bibat in any way (please do this!), the
-:ref:`contributing` section has some useful information.
+If you would like to contribute to bibat in any way (please do!), the
+:ref:`contributing` page has some useful information.
 
 If you are looking for reference information about bibat's :ref:`cli` or
-:ref:`api`, check those out!
+:ref:`api`, check out the :ref:`reference` page!
 
-.. _installation:
+.. toctree::
+   :maxdepth: 2
+
+   working_with_bibat
+   contributing
+   reference
+
+.. _getting_started:
+
+Getting started
+===============
 
 Installation
-============
+------------
 
 You can install bibat like this (make sure you are in a Python environment where
 you would like to install bibat):
@@ -44,12 +49,9 @@ To install the latest version of bibat from github:
 .. code:: sh
 
     $ pip install git+https://github.com/teddygroves/bibat.git@main
-    
-
-.. _usage:
 
 Usage
-=====
+-----
 
 Bibat is intended to be used from the command line like this:
 
@@ -78,434 +80,3 @@ See `bibat's cookiecutter schema
 <https://github.com/teddygroves/bibat/blob/main/bibat/cookiecutter.json>`_ for
 the information you need to create a suitable json file. You can use bibat to
 start your statistical analysis project like this:
-
-.. _intended workflow:
-
-Intended workflow
-=================
-
-bibat assumes that a statistical analysis consists of the following components:
-
-- **Raw data**
-- **Data preparation** in which raw data is transformed to produce prepared data
-  that fit a common structure.
-- **Statistical models**
-- **Inferences**, i.e. combinations of a prepared dataset, a statistical model,
-  a choice of which modes to fit the model in (for example
-  prior mode, posterior mode and exact k-fold cross-validation mode) and how
-  exactly to do so, and the results of doing this fitting.
-- **Investigations** that do things with inferences and prepared data, such as
-  making plots.
-- **Documentation**
-
-These terms are all pretty standard, except "Inference", which has a more
-specific than normal meaning in this context. This choice was made to avoid
-creating new jargon, and in order to match the arviz concept of `InferenceData
-<https://arviz-devs.github.io/arviz/api/inference_data.html>`_.
-  
-To perform a statistical analysis means running some data preparation
-operations, then carrying out a set of inferences, doing some investigations
-and providing some documentation. bibat provides functionality for conveniently
[…]
