+++ bibat/{{cookiecutter.repo_name}}/docs/index.rst
@@ -3,18 +3,57 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to project_name's documentation!
-========================================
+Welcome to {{ cookiecutter.project_name }}'s documentation!
+===========================================================
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
+  how_to_build
+  how_to_do_some_things_with_rst
+  reference
+
+
 Write your documentation here in `reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_.
+
+.. _how_to_build:
+
+How to do build the documentation
+=================================
+
+To create an html version of this documentation, run the command
+
+.. code:: sh
+
+   make docs
+
+from the project root directory.
+
+
+.. _how_to_do_some_things_with_rst:
+
+How to do some things with reStructuredText
+===========================================
 
 Include images like this:
 
 .. image:: img/example.png
+
+
+.. _reference:
+
+Reference
+=========
+
+The main reason to use Sphinx is for automatic documentation of Python code.
+
+Data preparation module
+-----------------------
+
+.. automodule:: {{ cookiecutter.repo_name }}.data_preparation
+   :members:
+
 
 Indices and tables
 ==================
