+++ pyproject.toml
@@ -55,26 +55,20 @@ bibat = "bibat.cli:generate_project"
 [tool.black]
 line-length = 80
 target-version = ['py311']
-include = '\.pyi?$'
-exclude = '''
-
-(
-  /(
-      \.eggs         # exclude a few common directories in the
-    | \.git          # root of the project
-    | \.hg
-    | \.mypy_cache
-    | \.tox
-    | \.venv
-    | _build
-    | buck-out
-    | build
-    | dist
-    | bibat/{{cookiecutter.repo_name}}
-  )/
-  | foo.py           # also separately exclude a file named foo.py in
-                     # the root of the project
-)
+extend-exclude = '''
+/(
+   \.eggs         # exclude a few common directories in the
+  | \.git          # root of the project
+  | \.hg
+  | \.mypy_cache
+  | \.tox
+  | \.venv
+  | _build
+  | buck-out
+  | build
+  | dist
+  | bibat/\{\{cookiecutter\.repo_name\}\}
+)/
 '''
 
 [tool.isort]
