+++ .pre-commit-config.yaml
@@ -9,6 +9,7 @@ repos:
     rev: 22.10.0
     hooks:
     -   id: black
+        exclude: bibat/\{\{cookiecutter\.repo_name\}\}
 -   repo: https://github.com/pycqa/isort
     rev: 5.12.0
     hooks:
