### 2023-05-04 7fe0a08
fix pyproject toml pytest typo
