+++ bibat/{{cookiecutter.repo_name}}/pyproject.toml
@@ -22,7 +22,7 @@ dependencies = [
     "scipy",
     "scikit-learn",
     "toml",
-    {-% if cookiecutter.create_tests_directory == 'y' 0%}    "pytest",
+    {-% if cookiecutter.create_tests_directory == 'y' -%}    "pytest",
     "black",{-% endif -%}
     {-% if cookiecutter.docs_format == 'Sphinx' -%}"sphinx",{-% endif -%}
 ]
