+++ README.rst
@@ -8,29 +8,29 @@ Bibat: Batteries-Included Bayesian Analysis Template
 .. image:: https://readthedocs.org/projects/bibat/badge/?version=latest
     :target: https://bibat.readthedocs.io/en/latest/?badge=latest
     :alt: Documentation Status
-    
+
 .. image:: https://github.com/teddygroves/bibat/actions/workflows/run_tox.yml/badge.svg
     :target: https://github.com/teddygroves/bibat/actions/workflows/run_tox.yml
     :alt: Test
-    
+
 .. image:: https://github.com/teddygroves/bibat/actions/workflows/test_end_to_end.yml/badge.svg
     :target: https://github.com/teddygroves/bibat/actions/workflows/test_end_to_end.yml
     :alt: Test
-    
+
 .. image:: https://www.repostatus.org/badges/latest/active.svg
    :alt: Project Status: Active – The project has reached a stable, usable state and is being actively developed.
    :target: https://www.repostatus.org/#active
- 
+
 .. image:: https://img.shields.io/badge/python->=3.9-blue.svg
    :alt: Supported Python versions: 3.9 and newer
    :target: https://www.python.org/
-   
+
 .. image:: https://badge.fury.io/py/bibat.svg
     :target: https://badge.fury.io/py/bibat
 
 
 Bibat is a Python package providing a flexible interactive template for Bayesian
-statistical analysis projects. 
+statistical analysis projects.
 
 It aims to make it easier to create software projects that implement a Bayesian
 workflow that scales to arbitrarily many inter-related statistical models, data
@@ -71,7 +71,7 @@ To install the latest version of bibat from github:
 .. code:: sh
 
     $ pip install git+https://github.com/teddygroves/bibat.git@main
-    
+
 
 To install bibat with development dependencies:
 
@@ -109,6 +109,93 @@ See `bibat's cookiecutter schema
 <https://github.com/teddygroves/bibat/blob/main/bibat/cookiecutter.json>`_ for
 the information you need to create a suitable json file.
 
+Dependencies
+============
+
+Bibat's dependencies fall into two categories: a few that are required by bibat
+itself, and some more that are required in order to run the analyses that bibat
+generates.
+
+Dependencies required by bibat
+------------------------------
+
+Bibat requires Python version 3.9 or greater.
+
+Bibat's other dependencies are all Python packages. These can be found in
+bibat's `pyproject.toml file
+<https://github.com/teddygroves/bibat/blob/main/pyproject.toml>`_.
+
+The following packages are required in order to run bibat and are installed
+automatically when you run :code:`pip install bibat`:
+
+- cookiecutter
+- click
+- pydantic
+
+
+Bibat also has the following development dependencies, which can be installed
+by running :code:`pip install bibat'[development]'`:
+
+- black
+- isort
+- pre-commit
+- pytest
+- tox
+- sphinx
+- sphinx-click
+- pydata_sphinx_theme
+
+
+Target project dependencies: Python
+-----------------------------------
+
+Projects generated by bibat have their Python dependencies listed in the file
+:code:`pyproject.toml`.  The cookiecutter template that generates this file can
+be found `here
+<https://github.com/teddygroves/bibat/blob/main/bibat/%7B%7Bcookiecutter.repo_name%7D%7D/pyproject.toml>`_.
+These packages will be installed when you run :code:`make env` or `make
+analysis` and are as follows:
+
+- arviz
+- cmdstanpy
+- jupyter
+- numpy
+- pandas
+- pandera
+- pydantic
+- scipy
+- scikit-learn
+- toml
+
+In addition, the following Python packages may be installed, depending on how
+the user answers bibat's wizard:
+
+- pytest
+- black
+- sphinx
+
[…]
