+++ .github/workflows/coverage.yml
@@ -1,0 +1,8 @@
+name: Generate test coverage report
+on: push
+jobs:
+  codecov:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Upload coverage reports to Codecov
+        uses: codecov/codecov-action@v3
