+++ .github/workflows/run_tox.yml
@@ -25,3 +25,5 @@ jobs:
         pip install -e .[development]
     - name: Run tox
       run: tox
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v3
