+++ tox.ini
@@ -1,4 +1,5 @@
 [tox]
+isolated_build = True
 env_list =
     format
     flake8
@@ -15,16 +16,18 @@ package = wheel
 wheel_build_env = .pkg
 deps =
     pytest>=7
+    codecov
+    pytest-cov
     pytest-sugar
 commands =
-    pytest {tty:--color=yes} tests
+    pytest {tty:--color=yes} tests --cov=bibat --cov-report xml --cov-report term
 
 [testenv:format]
 description = install black and isort and invoke them on the current folder
 deps =
     black[jupyter]
     isort
-skip_install = true
+skip_install = True
 commands =
     black --check --diff .
     isort --check-only --diff .
