+++ pyproject.toml
@@ -34,7 +34,9 @@ development = [
     "black",
     "isort",
     "pre-commit",
+    "codecov",
     "pytest",
+    "pytest-cov",
     "tox",
     "sphinx",
     "sphinx-click",
