+++ tox.ini
@@ -1,5 +1,4 @@
 [tox]
-isolated_build = True
 env_list =
     format
     flake8
@@ -12,6 +11,7 @@ max_line_length = 80
 
 [testenv]
 description = run the tests with pytest
+skip_install = true
 package = wheel
 wheel_build_env = .pkg
 deps =
@@ -20,7 +20,8 @@ deps =
     pytest-cov
     pytest-sugar
 commands =
-    pytest {tty:--color=yes} tests --cov=bibat --cov-report xml --cov-report term
+    pip install -e .'[development]'
+    pytest tests --cov --cov-report xml --cov-report term
 
 [testenv:format]
 description = install black and isort and invoke them on the current folder
