+++ README.rst
@@ -28,6 +28,8 @@ Bibat: Batteries-Included Bayesian Analysis Template
 .. image:: https://badge.fury.io/py/bibat.svg
     :target: https://badge.fury.io/py/bibat
 
+.. image:: https://codecov.io/github/teddygroves/bibat/branch/main/graph/badge.svg?token=ck0IKyzP7J
+    :target: https://codecov.io/github/teddygroves/bibat
 
 Bibat is a Python package providing a flexible interactive template for Bayesian
 statistical analysis projects.
