### 2023-05-08 23e8fc8
remove codecov action as tox should make it unnecessary
