### 2023-05-08 209592b
Add unit test for cli functions
