+++ tests/test_unit/test_cli.py
@@ -1,3 +1,9 @@
+"""Unit tests for the cli module.
+
+Adapted from examples here: https://click.palletsprojects.com/en/8.1.x/testing/
+
+"""
+
 import os
 import shutil
 
@@ -7,6 +13,7 @@ from bibat.cli import generate_project
 
 
 def test_generate_project():
+    """Test that the generate_project function works when called normally."""
     runner = CliRunner()
     result = runner.invoke(generate_project)
     assert result.exit_code == 0
@@ -14,6 +21,7 @@ def test_generate_project():
 
 
 def test_generate_project_with_config():
+    """Test that the generate_project function works with a config file."""
     runner = CliRunner()
     f = os.path.join("tests", "data", "example_config.yml")
     result = runner.invoke(generate_project, ["--config-file", f])
