+++ .github/workflows/test_end_to_end.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
         matrix:
             os: [ubuntu-latest, windows-latest]
-
+            python-version: ["3.9", "3.10", "3.11"]
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python 3.11.3
