+++ .github/workflows/test_end_to_end.yml
@@ -15,10 +15,10 @@ jobs:
             python-version: ["3.9", "3.10", "3.11"]
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python 3.11.3
+    - name: Set up Python
       uses: actions/setup-python@v2
       with:
-        python-version: 3.11.3
+        python-version: ${{ matrix.python-version }}
     - name: Install bibat
       run: pip install -e .
     - name: Use bibat
