### 2023-05-08 8c2eb52
Add test for some bad wizard cases
