### 2023-05-08 f58869c
Correct wizard tests
