### 2023-05-23 973b00d
delete unused global variables in data preparation file
