+++ bibat/examples/baseball/baseball/data_preparation.py
@@ -9,23 +9,14 @@ import os
 
 import pandas as pd
 import pandera as pa
-from baseball.util import CoordDict
 from pandera.typing import DataFrame, Series
 from pydantic.dataclasses import dataclass
 
+from baseball.util import CoordDict
+
 NAME_FILE = "name.txt"
 COORDS_FILE = "coords.json"
-MEASUREMENTS_FILE = "measurements.csv"
-NEW_COLNAMES = {"yButIThoughtIdAddSomeLetters": "y"}
-DROPNA_COLS = ["y"]
 N_CV_FOLDS = 10
-DIMS = {
-    "b": ["covariate"],
-    "y": ["observation"],
-    "yrep": ["observation"],
-    "llik": ["observation"],
-}
-
 HERE = os.path.dirname(__file__)
 DATA_DIR = os.path.join(HERE, "..", "data")
 RAW_DIR = os.path.join(DATA_DIR, "raw")
