### 2023-05-23 6d66ed0
Make rst links in README.md anonymous
