### 2023-05-23 ba0fb3b
Fix mistakes in baseball vignette
