### 2023-05-23 ee7eff8
Fix Makefile whitespace in case with no tests
