### 2023-05-31 220f442
Refactor to add an explicit fitting mode module
