+++ bibat/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/fitting_mode.py
@@ -1,0 +1,130 @@
+"""A general definition of a fitting mode, plus some mode instances."""
+
+import textwrap
+from typing import Callable, Dict, Union
+
+import numpy as np
+import xarray as xr
+from cmdstanpy import CmdStanMCMC, CmdStanModel
+from pydantic import BaseModel
+from sklearn.model_selection import KFold
+
+
+class FittingMode(BaseModel):
+    """A way of fitting a statistical model.
+
+    Each fitting mode has a name, an 'idata_target', i.e. the group that this
+    mode will create in the output InferenceData, and a function that performs
+    the fitting.
+
+    Inferences can be configured to use any of the fitting modes defined here
+    by including them by name in the top-level list 'modes'. For example:
+
+      ...
+      modes = ['prior', 'posterior', 'kfold']
+      ...
+
+    The idata_target can be any name you want; however if it is 'prior' or
+    'posterior' it must be compatible with how this package's 'sample' module
+    creates groups with these names.
+
+    Each fitting mode's fit function must match the signature specified in the
+    FittingMode class. This can also be changed, but must agree with the
+    'sample' module.
+
+
+    """
+    name: str
+    idata_target: str
+    fit: Callable[[CmdStanModel, Dict, Dict[str, str]], Union[CmdStanMCMC, xr.Dataset]]
+
+
+def fit_prior(model: CmdStanModel, input_dict: dict, kwargs) -> CmdStanMCMC:
+    """Create a CmdStanMCMC from a model, data and config, in prior mode.
+
+    :param model: a CmdStanModel. It must have a data variable called
+    'likelihood': setting this variable to 0 should trigger prior mode,
+    typically by causing the code that performs likelihood evaluations not to
+    run.
+
+    :param input_dict: a Stan input dictionary. It doesn't need to have a
+    'likelihood' entry: if it does, it will be overwritten.
+
+    :param kwargs: keyword arguments for CmdStanModel.sample
+
+    """
+    input_dict_final = input_dict | {"likelihood": 0}
+    return model.sample(input_dict_final, **kwargs)
+
+
+def fit_posterior(model: CmdStanModel, input_dict: dict, kwargs) -> CmdStanMCMC:
+    """Create a CmdStanMCMC from a model, data and config, in posterior mode.
+
+    :param model: a CmdStanModel. It must have a data variable called
+    'likelihood': setting this variable to 1 should trigger posterior mode,
+    typically by causing the code that performs likelihood evaluations to run.
+
+    :param input_dict: a Stan input dictionary. It doesn't need to have a
+    'likelihood' entry: if it does, it will be overwritten.
+
+    :param kwargs: keyword arguments for CmdStanModel.sample
+    """
+    input_dict_final = input_dict | {"likelihood": 1}
+    return model.sample(input_dict_final, **kwargs)
+
+
+def fit_kfold(model: CmdStanModel, input_dict :dict, kwargs) -> xr.DataArray:
+    """Do k-fold cross validation, given a CmdStanModel, some data and config.
+
+    :param model: a CmdStanModel. It must have a data variables called
+    'likelihood', 'N_train', 'N_test', 'ix_train' and 'ix_test'.
+
+    :param input_dict: a Stan input dictionary. It doesn't need to have any of
+    the required variables set (they will be overwritten if they are set).
+
+    :param kwargs: dictionary with an entry for 'n_folds' that specifies the
+    value of k for k-fold cross-validation, plus keyword arguments for
+    CmdStanModel.sample
+
+    """
+    if "n_folds" not in kwargs.keys():
+        msg = textwrap.dedent(
+            """
[…]
