### 2023-05-31 684b638
Fix log likelihood group assignment
