+++ bibat/{{cookiecutter.repo_name}}/Makefile
@@ -5,7 +5,8 @@ ACTIVATE_VENV = .venv/bin/activate
 SRC = {{cookiecutter.repo_name}}
 {% if cookiecutter.docs_format != 'No docs' %}DOCS_DIR = docs{% endif %}
 {% if cookiecutter.docs_format == 'Sphinx' %}DOCS_BUILDDIR = docs/build{% endif %}
-{% if cookiecutter.docs_format == 'Quarto' %}REPORT_STEM = docs/report{% endif %}
+{% if cookiecutter.docs_format == 'Quarto' %}REPORT_STEM = docs/report
+QUARTO_EXTENSIONS_FOLDER = $(DOCS_DIR)/_extensions{% endif %}
 
 ifeq ($(OS),Windows_NT)
 	INSTALL_CMDSTAN_FLAGS = --compiler
@@ -17,10 +18,15 @@ endif
 $(ACTIVATE_VENV):
 	python -m venv .venv --prompt={{cookiecutter.repo_name}}
 
-{% if cookiecutter.docs_format != 'No docs' %}docs: $(ENV_MARKER)
-	. $(ACTIVATE_VENV) && (\{% endif %}
+{% if cookiecutter.docs_format == 'Quarto' -%}
+$(QUARTO_EXTENSIONS_FOLDER):
+	cd docs && quarto add quarto-ext/include-code-files && cd -
+{%- endif %}
+
+{% if cookiecutter.docs_format != 'No docs' %}docs: $(ENV_MARKER) {%- if cookiecutter.docs_format == 'Quarto' %} $(QUARTO_EXTENSIONS_FOLDER){% endif %}
+	. $(ACTIVATE_VENV) && (\ {%- endif %}
 		{% if cookiecutter.docs_format == 'Sphinx' %}sphinx-build "$(DOCS_DIR)" "$(DOCS_BUILDDIR)" -b html; \
-    {%- elif cookiecutter.docs_format == 'Quarto' -%}quarto render $(REPORT_STEM).qmd; \
+   	{% elif cookiecutter.docs_format == 'Quarto' %}quarto render $(REPORT_STEM).qmd; \
 		{%- endif %}
 {% if cookiecutter.docs_format != 'No docs' %}	){% endif %}
 
@@ -34,7 +40,7 @@ $(ENV_MARKER): $(ACTIVATE_VENV) $(CMDSTAN)
 
 analysis: $(ENV_MARKER)
 	. $(ACTIVATE_VENV) && (\{%- if cookiecutter.create_tests_directory == 'y' %}
-	  python -m pytest || exit 1; \{% endif -%}
+	  python -m pytest || exit 1; \{%- endif %}
 	  python $(SRC)/prepare_data.py || exit 1; \
 	  python $(SRC)/sample.py || exit 1; \
 	  jupyter execute $(SRC)/investigate.ipynb || exit 1; \
