+++ bibat/examples/baseball/baseball/sample.py
@@ -1,81 +1,20 @@
-"""Run all the configs in the model_configurations folder."""
+"""Run all the inferences in the inferences folder."""
 
 import os
-from typing import Dict
 
 import arviz as az
 import cmdstanpy
-import numpy as np
-import xarray as xr
-from sklearn.model_selection import KFold
 
 from baseball.data_preparation import load_prepared_data
-from baseball.inference_configuration import (
-    AVAILABLE_MODES,
-    InferenceConfiguration,
-    load_inference_configuration,
-)
-from baseball.util import CoordDict
+from baseball.inference_configuration import load_inference_configuration
 
 HERE = os.path.dirname(__file__)
 RUNS_DIR = os.path.join(HERE, "..", "inferences")
 STAN_DIR = os.path.join(HERE, "stan")
 
 
-def get_sample_kwargs(raw: Dict, mode: str) -> Dict:
-    """Get the sampel kwargs for a mode from raw toml."""
-    out = {k: v for k, v in raw.items() if k not in AVAILABLE_MODES}
-    if mode in raw.keys():
-        out |= raw[mode]
-    return out
-
-
-def get_kfold_llik(
-    ic: InferenceConfiguration,
-    stan_input_base: Dict,
-    model: cmdstanpy.CmdStanModel,
-    coords: CoordDict,
-) -> xr.DataArray:
-    """Run k-fold cross-validation and return out-of-sample log likelihood.
-
-    :param ic: InferenceConfiguration, must have non-null field kfold_folds
-
-    :param stan_input_base: stan input dictionary
-
-    :param model: CmdStanModel
-    """
-    n_folds = ic.kfold_folds
-    sample_kwargs = get_sample_kwargs(ic.sample_kwargs, "kfold")
-    kf = KFold(n_folds, shuffle=True, random_state=1234)
-    lliks_by_fold = []
-    full_ix = np.array(stan_input_base["ix_train"])
-    for fold, (ix_train, ix_test) in enumerate(kf.split(full_ix)):
-        input_dict = stan_input_base | {
-            "likelihood": 1,
-            "N_train": len(ix_train),
-            "N_test": len(ix_test),
-            "ix_train": full_ix[ix_train].tolist(),
-            "ix_test": full_ix[ix_test].tolist(),
-        }
-        mcmc = model.sample(data=input_dict, **sample_kwargs)
-        llik_fold = mcmc.draws_xr(vars=["llik"])
-        # remember the fold
-        llik_fold["fold"] = fold
-        llik_fold = llik_fold.set_coords("fold")
-        # set value of index "chain" to zero to match arviz convention
-        llik_fold = llik_fold.assign_coords(
-            {"new_chain": ("chain", [0])}
-        ).set_index(chain="new_chain")
-        # get observation correct
-        llik_fold["llik_dim_0"] = np.array(coords["observation"])[ix_test]
-        # rename coord "llik_dim_0" to "observation"
-        llik_fold = llik_fold.rename({"llik_dim_0": "observation"})
-        lliks_by_fold.append(llik_fold["llik"])
-    return xr.concat(lliks_by_fold, dim="observation").sortby("observation")
-
-
 def main():
-    """Run the main script."""
+    """Fit all inferences in all modes."""
     run_dirs = [
         os.path.join(RUNS_DIR, d)
         for d in os.listdir(RUNS_DIR)
@@ -100,20 +39,27 @@ def main():
             "coords": prepared_data.coords,
             "dims": ic.dims,
         }
-        for mode, likelihood in [("prior", 0), ("posterior", 1)]:
-            if mode in ic.modes:
-                sample_kwargs = get_sample_kwargs(ic.sample_kwargs, mode)
-                input_dict = stan_input_base | {"likelihood": likelihood}
-                idata_kwargs[mode] = model.sample(
[…]
