+++ bibat/examples/baseball/inferences/normal2006/config.toml
@@ -14,3 +14,8 @@ warn-pedantic = true
 save_warmup = false
 iter_warmup = 2000
 iter_sampling = 2000
+
+[mode_options.prior]
+chains = 2
+iter_warmup = 1000
+iter_sampling = 1000
