+++ docs/cli.rst
@@ -1,0 +1,12 @@
+.. _cli:
+
+===
+CLI
+===
+
+This page documents bibat's command line interface, i.e. the command-line
+commands that bibat provides.
+
+.. click:: bibat.cli:generate_project
+   :prog: bibat
+   :nested: full
