+++ docs/reference.rst
@@ -1,45 +1,0 @@
-.. _reference:
-
-=========
-Reference
-=========
-
-This page has automatically-generated reference information that you can use to
-find out how to use bibat's :ref:`command line interface<cli>` and :ref:`python
-functions<API>`.
-
-.. _cli:
-
-CLI
-===
-
-This section lists the commands that bibat provides and shows how to use them.
-
-.. click:: bibat.cli:generate_project
-   :prog: bibat
-   :nested: full
-
-.. _api:
-
-Python API
-==========
-
-Although bibat is intended to be used from the command line, it is also
-possible to use it from Python. This section goes through the Python modules
-that bibat provides and shows how to use the provided functions.
-
-
-Wizarding module
-----------------
-
-.. automodule:: bibat.wizarding
-   :members:
-
-Cli module
-----------
-
-.. automodule:: bibat.cli
-   :members:
-   :undoc-members:
-
-   .. autofunction:: generate_project
