### 2023-06-02 c58e4b0
delete old CONTRIBUTING.md
