### 2023-06-05 ff58230
add lighthouse audit action and language in docs/conf.py
