+++ .github/workflows/test_sphinx.yml
@@ -6,7 +6,7 @@ on: push
 
 jobs:
   docs:
-    
+
     runs-on: ubuntu-latest
 
     steps:
@@ -22,3 +22,10 @@ jobs:
       run: |
         cd bibat/\{\{cookiecutter.repo_name\}\}
         sphinx-build docs ./docs/_build/html/
+    - name: Audit with Lighthouse
+      uses: treosh/lighthouse-ci-action@v10
+      with:
+        configPath: ".github/workflows/lighthouserc.json"
+        temporaryPublicStorage: true
+        uploadArtifacts: true
+        runs: 3 # Multiple runs to reduce variance
