### 2023-06-05 c191248
remove lighthouse section from target project docs check
