+++ .github/workflows/test_target_project_docs.yml
@@ -22,10 +22,3 @@ jobs:
       run: |
         cd bibat/\{\{cookiecutter.repo_name\}\}
         sphinx-build docs ./docs/_build/html/
-    - name: Audit with Lighthouse
-      uses: treosh/lighthouse-ci-action@v10
-      with:
-        configPath: ".github/workflows/lighthouserc.json"
-        temporaryPublicStorage: true
-        uploadArtifacts: true
-        runs: 3 # Multiple runs to reduce variance
