+++ docs/accessibility.rst
@@ -1,0 +1,44 @@
+.. _accessibility:
+
+=======================
+Accessibility statement
+=======================
+
+The bibat developers want as many people as possible to be able to use this
+website. For example, that means you should be able to:
+
+- change colours, contrast levels and fonts
+
+- zoom in up to 300% without the text spilling off the screen
+
+- navigate most of the website using just a keyboard
+
+- navigate most of the website using speech recognition software
+
+- listen to most of the website using a screen reader (including the most
+  recent versions of JAWS, NVDA and VoiceOver)
+
+- We’ve also made the website text as simple as possible to understand.
+
+How accessible this website is
+------------------------------
+
+We know some parts of this website are not fully accessible:
+
+-
+
+Feedback and contact information
+--------------------------------
+
+If you need information on this website in a different format like accessible
+PDF, large print, easy read, audio recording or braille, please contact us by
+`email <mailto:groves.teddy@gmail.com>`__ or raise an issue on the `bibat GitHub repository
+<https://github.com/teddygroves/bibat/>`__.
+
+Reporting accessibility problems with this website
+--------------------------------------------------
+
+We’re always looking to improve the accessibility of this website. If you find
+any problems not listed on this page or think we’re not meeting accessibility
+requirements, contact us by `email <mailto:groves.teddy@gmail.com>`__ or raise
+an issue on our `GitHub repository <https://github.com/teddygroves/bibat/>`__..
