+++ docs/index.rst
@@ -8,171 +8,30 @@ uses `cookiecutter <https://cookiecutter.readthedocs.io/en/latest/>`_, `Stan
 `arviz <https://arviz-devs.github.io/arviz/>`_. It takes a 'batteries-included'
 approach, creating a whole working project rather than a skeleton.
 
-This page has the information you need in order to :ref:`get started
-<getting_started>` with bibat by installing it and then using it to create a
-folder containing a complete example Bayesian statistical analysis project. It
-also describes bibat's :ref:`dependencies <dependencies>`.
+The :ref:`getting started <getting_started>` page has the information you need
+in order to install bibat it and then using it to create a folder containing a
+complete example Bayesian statistical analysis project. It also describes
+bibat's :ref:`dependencies <dependencies>`.
 
 The next thing you will probably want to do is learn how to turn the example
 analysis into the analysis you actually want to do. For this there is the
 :ref:`working with bibat <working_with_bibat>` page.
 
-If you would like to contribute to bibat in any way (please do!), the
-:ref:`contributing` page has some useful information.
-
 For reference information about bibat, check out the :ref:`cli` and :ref:`api`
 pages.
+
+If you would like to contribute to bibat in any way---please do!---the
+:ref:`contributing` page has some useful information.
+
+Finally, see the :ref:`accessibility` page to find out how we have tried to
+make this website as useable as possible for everyone, and for how to
 
 .. toctree::
    :maxdepth: 2
 
+   getting_started
    working_with_bibat
    cli
    api
    contributing
    accessibility
-
-.. _getting_started:
-
-Getting started
-===============
-
-Installation
-------------
-
-You can install bibat like this (make sure you are in a Python environment where
-you would like to install bibat):
-
-.. code:: sh
-
-    $ pip install bibat
-
-To install the latest version of bibat from GitHub:
-
-.. code:: sh
-
-    $ pip install git+https://github.com/teddygroves/bibat.git@main
-
-Usage
------
-
-Bibat is intended to be used from the command line like this:
-
-.. code:: shell
-
-    $ bibat
-
-Running this command will trigger a command line wizard. After following the
-wizard's instructions, a new directory will be created that implements a simple
-statistical analysis. To try out the example analysis, run the following
-command from the root of the new directory:
-
-.. code:: shell
-
-    $ make analysis
-
-If you already know how you are going to answer the wizard's questions, you can
-put your answers in a json file with relative path `my_json_file.json` and run
-bibat like this:
-
-.. code:: shell
-
-    $ bibat --config_file=my_json_file.json
-
-See `bibat's cookiecutter schema
-<https://github.com/teddygroves/bibat/blob/main/bibat/cookiecutter.json>`_ for
-the information you need to create a suitable json file. You can use bibat to
-start your statistical analysis project like this:
-
-
-.. _dependencies:
-
-Dependencies
-============
-
-Bibat's dependencies fall into two categories: a few that are required by bibat
-itself, and some more that are required in order to run the analyses that bibat
-generates.
-
-Dependencies required by bibat
-------------------------------
-
-Bibat requires Python version 3.9 or greater.
-
-Bibat's other dependencies are all Python packages. These can be found in
-bibat's `pyproject.toml file
-<https://github.com/teddygroves/bibat/blob/main/pyproject.toml>`_.
-
-The following packages are required in order to run bibat and are installed
-automatically when you run :code:`pip install bibat`:
-
-- cookiecutter
-- click
-- pydantic
-
-
[…]
