+++ docs/requirements.txt
@@ -1,0 +1,2 @@
+furo
+sphinx-click
