### 2023-06-22 6788e76
Fix typos in docs/index.rst
