+++ docs/index.rst
@@ -9,9 +9,9 @@ uses `cookiecutter <https://cookiecutter.readthedocs.io/en/latest/>`_, `Stan
 approach, creating a whole working project rather than a skeleton.
 
 The :ref:`getting started <getting_started>` page has the information you need
-in order to install bibat it and then using it to create a folder containing a
-complete example Bayesian statistical analysis project. It also describes
-bibat's :ref:`dependencies <dependencies>`.
+in order to install bibat it and then use it to create a complete example
+Bayesian statistical analysis project. It also describes bibat's
+:ref:`dependencies <dependencies>`.
 
 The next thing you will probably want to do is learn how to turn the example
 analysis into the analysis you actually want to do. For this there is the
@@ -20,8 +20,8 @@ analysis into the analysis you actually want to do. For this there is the
 For reference information about bibat, check out the :ref:`cli` and :ref:`api`
 pages.
 
-If you would like to contribute to bibat in any way---please do!---the
-:ref:`contributing` page has some useful information.
+If you would like to contribute to bibat in any way, the :ref:`contributing`
+page has some useful information.
 
 Finally, see the :ref:`accessibility` page to find out how we have tried to
 make this website as useable as possible for everyone, and for how to
