### 2023-06-22 b05c605
Fix typos in docs/index.rst
