### 2023-06-22 b720a9b
Fix incorrect file in contributing docs
