### 2023-06-22 24c0c75
Remove minor python versions from github actions recipes
