+++ .github/workflows/release.yml
@@ -14,7 +14,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v1
       with:
-        python-version: 3.11.3
+        python-version: 3.11
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
