+++ .github/workflows/test_docs.yml
@@ -13,10 +13,10 @@ jobs:
     steps:
 
     - uses: actions/checkout@v2
-    - name: Set up Python 3.11.3
+    - name: Set up Python 3.11
       uses: actions/setup-python@v2
       with:
-        python-version: 3.11.3
+        python-version: 3.11
     - name: install bibat with development packages
       run: pip install -e .\[development\]
     - name: build docs with sphinx
