+++ README.rst
@@ -157,8 +157,8 @@ Projects generated by bibat have their Python dependencies listed in the file
 :code:`pyproject.toml`.  The cookiecutter template that generates this file can
 be found `here
 <https://github.com/teddygroves/bibat/blob/main/bibat/%7B%7Bcookiecutter.repo_name%7D%7D/pyproject.toml>`_.
-These packages will be installed when you run :code:`make env` or `make
-analysis` and are as follows:
+These packages will be installed when you run :code:`make env` or :code:`make analysis`
+and are as follows:
 
 - arviz
 - cmdstanpy
@@ -183,8 +183,8 @@ Target project dependencies: Cmdstan
 
 Bibat will attempt to install `cmdstan
 <https://mc-stan.org/docs/cmdstan-guide/index.html>`__, the command line
-interface to Stan, when you run the commands :code:`make env` or :code:`make
-analysis` from the root of the target project.
+interface to Stan, when you run the commands :code:`make env` or :code:`make analysis` 
+from the root of the target project.
 
 If bibat fails to install cmdstan, please raise an issue! The relevant
 parts of the `cmdstan
