### 2023-07-07 59fb253
update target project for pydantic v2
