+++ bibat/examples/baseball/docs/report.qmd
@@ -212,7 +212,7 @@ The next step is to write functions that return `PreparedData` objects. In this
 case I wrote a couple of data preparation functions: `prepare_data_2006` and
 `prepare_data_bdb`:
 
-```{.python include="../baseball/data_preparation.py" start-line=77 end-line=161}
+```{.python include="../baseball/data_preparation.py" start-line=78 end-line=162}
 ```
 
 To take into account the inconsistency between the two raw data sources, I
@@ -224,7 +224,7 @@ first had to change the variable `RAW_DATA_FILES`:
 Next I changed the `prepare_data` function to handle the two different data
 sources.
 
-```{.python include="../baseball/data_preparation.py" start-line=35 end-line=54}
+```{.python include="../baseball/data_preparation.py" start-line=36 end-line=56}
 ```
 
 To finish off I deleted the unused global variables `NEW_COLNAMES`,
