+++ paper.md
@@ -24,41 +24,45 @@ bibliography: bibliography.bib
 ---
 
 # Summary
-Bibat is a Python package providing a flexible interactive template for Bayesian
-statistical analysis projects. 
 
-It aims to make it easier to create software projects that implement a Bayesian
-workflow that scales to arbitrarily many inter-related statistical models, data
-transformations, inferences and computations. Bibat also aims to promote
-software quality by providing a modular, automated and reproducible project that
-takes advantage of and integrates together the most up to date statistical
-software.
+Bibat is a Python package that provides an interactive template for Bayesian
+statistical analysis projects.
 
-Bibat comes with "batteries included" in the sense that it creates a working
-example project, which the user can adapt so that it implements their desired
-analysis. We believe this style of template makes for better usability and
-easier testing of Bayesian workflow projects compared with the alternative
-approach of providing an incomplete skeleton project.
+Bibat aims to make it easier to create software projects that implement a
+Bayesian workflow that scales to arbitrarily many statistical models, data
+transformations and fitting methods. It also aims to promote software
+quality by providing a modular, automated and reproducible project that follows
+Python programming best practices and integrates together many popular
+statistical programming tools.
+
+Bibat is "batteries included" in the sense that it creates a working example
+project which the user can adapt so that it implements their desired analysis.
+This style of template makes for better usability and easier testing of
+Bayesian workflow projects compared with the alternative approach of providing
+an incomplete skeleton project.
 
 # Statement of need
-[@gelmanBayesianWorkflow2020] makes a compelling argument that, both in theory
-and practice, statisticians should explicitly recognise that applied data
-analysis projects typically involve many inter-related statistical models, data
-transformations, inferences and computations.
 
-However, there is currently little software that addresses Bayesian data
-analysis at the level of workflows. As a result, software that implements a
-Bayesian workflow for the purposes of a particular analysis is largely written
-by hand from scratch, following the design choices of an individual practitioner
-or team at the time they wrote the project. This practice wastes effort as much
-work is duplicated, and leads to sub-optimal software quality as expertise is
-not shared between practitioners (or between the same practitioner at different
-times).
+There is a growing interest in the concept of Bayesian workflow, as shown by
+papers like [@gelmanBayesianWorkflow2020],
+[@grinsztajnBayesianWorkflowDisease2021] and
+[@gabryVisualizationBayesianWorkflow2019]. There are software tools that
+address most individual aspects of a Bayesian workflow, from fetching and
+manipulating data to specifying, computing, storing, inspecting and documenting
+statistical inferences. However, there are relatively few software tools that
+address the problem of implementing a whole Bayesian workflow.
 
-Interactive project templates are an obvious solution to this problem, as they
-are already commonly used by software developers and machine learning
-practitioners to avoid wasted effort and share best practices when creating
-software projects that implement complex workflows. 
+This issue is especially pronounced for non-trivial Bayesian workflows
+involving many interacting, polytomous components (e.g. datasets, statistical
+models, diagnostics). Software that implements such workflows is difficult to
+write, and resources teaching how to do so are scarce, as pedagogical examples
+typically focus on simple cases.
+
+Interactive project templates are commonly used to address analogous issues
+in software engineering and machine learning. Compared with writing software
+from scratch, using a template is usually faster, promotes better code quality
+and provides a focal point for practitioners to share and improve best
+practices.
 
 There is currently there is no popular interactive project template that
[…]
