### 2023-07-14 d6e0e43
Temporary fix: use pydantic-v2 pandera branch
