+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "pip >= 20",
     "cookiecutter",
     "click",
-    "pydantic",
+    "pydantic >= 2.0.0",
 ]
 
 [project.optional-dependencies]
