+++ bibat/{{cookiecutter.repo_name}}/pyproject.toml
@@ -17,8 +17,8 @@ dependencies = [
     "jupyter",
     "numpy",
     "pandas",
-    "pandera",
-    "pydantic",
+    "pandera@git+https://github.com/unionai-oss/pandera.git@pydantic-v2",
+    "pydantic >= 2.0.0",
     "scipy",
     "scikit-learn",
     "toml",
