### 2023-07-14 7da1010
accomodate missing pandera dependency
